Factor out more route modules.
