+++ frontend/lib/data-requests/routes.ts
@@ -1,0 +1,9 @@
+import { ROUTE_PREFIX } from "../util/route-util";
+
+export function createDataRequestsRouteInfo(prefix: string) {
+  return {
+    [ROUTE_PREFIX]: prefix,
+    home: prefix,
+    multiLandlord: `${prefix}/multi-landlord`,
+  };
+}
