+++ frontend/lib/issues/issue-pages.tsx
@@ -9,7 +9,7 @@ import {
   toDjangoChoices,
 } from "../common-data";
 import Page from "../ui/page";
-import { IssuesRouteInfo, IssuesRouteAreaProps } from "../justfix-routes";
+import { IssuesRouteInfo, IssuesRouteAreaProps } from "./routes";
 import { Switch, Route } from "react-router";
 import { Link } from "react-router-dom";
 import { NotFound } from "../pages/not-found";
