+++ frontend/lib/justfix-cross-site-visitor-routes.ts
@@ -1,0 +1,9 @@
+export type JustfixCrossSiteVisitorRouteInfo = ReturnType<
+  typeof createJustfixCrossSiteVisitorRoutes
+>;
+
+export function createJustfixCrossSiteVisitorRoutes(prefix: string) {
+  return {
+    crossSiteAgreeToTerms: `${prefix}/terms`,
+  };
+}
