+++ frontend/lib/justfix-cross-site-visitor-steps.tsx
@@ -1,6 +1,6 @@
-import { JustfixCrossSiteVisitorRouteInfo } from "./justfix-routes";
 import { ProgressStepRoute } from "./progress/progress-step-route";
 import { createCrossSiteAgreeToTermsStep } from "./pages/cross-site-terms-opt-in";
+import { JustfixCrossSiteVisitorRouteInfo } from "./justfix-cross-site-visitor-routes";
 
 export function createJustfixCrossSiteVisitorSteps(
   routes: JustfixCrossSiteVisitorRouteInfo
